+++ src/views/Admin/CreateProduct.vue
@@ -102,7 +102,6 @@ export default {
         organization: db.collection('organizations').doc(department.organization.id),
         team: team.map(({ id }) => db.collection('users').doc(id)),
         archived: false,
-        slack: [],
       };
 
       try {